+++ package.json
@@ -10,7 +10,7 @@
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "styled-components": "^5.2.1",
-    "web-vitals": "^1.1.0",
+    "web-vitals": "^1.1.1",
     "yup": "^0.32.9"
   },
   "scripts": {
