+++ src/pages/index.js
@@ -1,5 +1,5 @@
 export { default as Home } from './home';
 export { default as About } from './about';
-export { default as Register } from './RegistrationForm/register';
-export { default as Post } from './post/Post';
+export { default as Register } from './register';
+export { default as Post } from './post';
 export { default as IndividualPost } from './post/IndividualPost';
