+++ src/pages/post/index.js
