+++ src/pages/register/index.js
